fix: fix lpdb files
