+++ src/app.feature/launchConjunctions/components/LPDBTable.tsx
@@ -1,63 +1,34 @@
 import { LPDBResponseDataType, LPDBResponseType } from '@app.modules/types/launchConjunctions'
 import React, { useEffect, useMemo, useRef, useState } from 'react'
 import styled from 'styled-components'
-import { Column, useTable, CellProps } from 'react-table'
+import { useTable, CellProps } from 'react-table'
 import { Table } from '@app.components/common/Table'
 import { useModal } from '@app.modules/hooks/useModal'
 import { useMutationDeleteLPDB } from '../query/useMutationLPDB'
-import { useQueryGetLPDBDownload } from '../query/useQueryLPDB'
 import CesiumModule from '@app.modules/cesium/cesiumModule'
-import { QueryObserverResult, RefetchOptions, RefetchQueryFilters } from 'react-query'
 import moment from 'moment'
 import LPDBDetailTable from './LPDBDetailTable'
-
-const COLUMNS: Column<LPDBResponseDataType>[] = [
-  {
-    Header: 'ID',
-    accessor: 'email',
-  },
-  {
-    Header: 'Type',
-    Cell: <div>LCA</div>,
-  },
-  {
-    Header: 'Upload Date',
-    accessor: 'createdAt',
-  },
-  {
-    Header: 'Status',
-    accessor: 'status',
-  },
-]
 
 type LPDBProps = {
   LPDBData: LPDBResponseDataType[]
   handleNewLaunchClick: () => void
   cesiumModule: CesiumModule
-  refetchLPDBData: <TPageData>(
-    options?: RefetchOptions & RefetchQueryFilters<TPageData>
-  ) => Promise<QueryObserverResult<LPDBResponseType, unknown>>
 }
 
-const LPDBTable = ({
-  LPDBData,
-  handleNewLaunchClick,
-  cesiumModule,
-  refetchLPDBData,
-}: LPDBProps) => {
+const LPDBTable = ({ LPDBData, handleNewLaunchClick, cesiumModule }: LPDBProps) => {
   const [isDoneStatusClicked, setIsDoneStatusClicked] = useState<boolean>(false)
   const [selectedLPDBId, setSelectedLPDBId] = useState<string>('')
-  const [selectedPath, setSelectedPath] = useState<string>('')
   const [trajectoryPath, setTrajectoryPath] = useState<string>('')
-  const columns = useMemo(() => COLUMNS, [])
+  // const { columns, data } = useTableData() -> useQuery로 이 안에서  해결
+
   const data = useMemo(() => LPDBData, [LPDBData])
   const tableContainerRef = useRef<HTMLDivElement>(null)
   const tableRef = useRef<HTMLTableElement>(null)
+
   const { modalType, modalVisible } = useModal('LAUNCHCONJUNCTIONS')
   const isLaunchConjunctionsClicked = modalType === 'LAUNCHCONJUNCTIONS' && modalVisible
-  const { mutate } = useMutationDeleteLPDB()
 
-  const { data: downloadData, refetch } = useQueryGetLPDBDownload(selectedPath)
+  const { mutate } = useMutationDeleteLPDB()
 
   const handleDetailClick = async (id: string, trajectoryPath) => {
     setSelectedLPDBId(id)
@@ -73,88 +44,64 @@ const LPDBTable = ({
     setIsDoneStatusClicked(false)
   }
 
-  const handleDownload = async (filePath: string) => {
-    setSelectedPath(filePath)
-    if (selectedPath) {
-      const response = await refetch()
-      const element = document.createElement('a')
-      const textFile = new Blob([response.data.data], {
-        type: 'text/plain',
-      })
-      element.href = URL.createObjectURL(textFile)
-      element.download = 'trajectory_result.txt'
-      document.body.appendChild(element)
-      element.click()
-      setSelectedPath('')
-    } else {
-      console.log('Selected path is empty...')
-    }
-  }
-
-  const { getTableProps, getTableBodyProps, headerGroups, prepareRow, rows } = useTable(
-    {
-      columns,
-      data,
-    },
-    (hooks) => {
-      hooks.visibleColumns.push((columns) => {
-        const newColumns = columns.map((columnObject) => {
-          if (columnObject.id === 'status') {
-            return {
-              ...columnObject,
-              Cell: ({ row, value }: CellProps<any>) => (
-                <>
-                  {value === 'DONE' ? (
-                    <div
-                      onClick={() => {
-                        handleDetailClick(row.original['_id'], row.original.trajectoryPath)
-                      }}
-                      style={{ color: '#fccb16', cursor: 'pointer' }}
-                    >
-                      <span>
-                        {'Success'}
-                        {'   '}
-                        <img
-                          src="/svg/right-arrow.svg"
-                          style={{ width: '5px', cursor: 'pointer' }}
-                          onClick={() => {}}
-                        />
-                      </span>
-                    </div>
-                  ) : (
-                    <div>{value === 'PENDING' ? 'Pending' : 'Failed'}</div>
-                  )}
-                </>
-              ),
-            }
-          }
-          if (columnObject.id === 'createdAt') {
-            return {
-              ...columnObject,
-              Cell: ({ value }: CellProps<any>) => (
-                <>{moment.utc(value).format('MMM DD, YY HH:mm:ss')}</>
-              ),
-            }
-          }
-          return columnObject
-        })
-        return [
-          ...newColumns,
-          {
-            id: 'delete',
-            Header: '',
-            Cell: ({ row }) => (
-              <img
-                src="/svg/trashbin.svg"
-                style={{ width: '14px', cursor: 'pointer' }}
-                onClick={() => handleDelete(row.original._id)}
-              />
-            ),
-          },
-        ]
-      })
-    }
+  const columns = useMemo(
+    () => [
+      {
+        Header: 'ID',
+        accessor: 'email',
+      },
+      {
+        Header: 'Type',
+        Cell: <div>LCA</div>,
+      },
+      {
+        Header: 'Upload Date',
+        accessor: 'createdAt',
+        Cell: ({ value }: CellProps<any>) => <>{moment.utc(value).format('MMM DD, YY HH:mm:ss')}</>,
+      },
+      {
+        Header: 'Status',
+        accessor: 'status',
+        Cell: ({ row, value }: CellProps<any>) => (
+          <>
+            {value === 'DONE' ? (
+              <div
+                onClick={() => {
+                  handleDetailClick(row.original['_id'], row.original.trajectoryPath)
+                }}
+                style={{ color: '#fccb16', cursor: 'pointer' }}
+              >
+                <span>
+                  {'Success'}
+                  {'   '}
+                  <img src="/svg/right-arrow.svg" style={{ width: '5px', cursor: 'pointer' }} />
+                </span>
+              </div>
+            ) : (
+              <div>{value === 'PENDING' ? 'Pending' : 'Failed'}</div>
+            )}
+          </>
+        ),
+      },
+      {
+        id: 'delete',
+        Header: '',
+        Cell: ({ row }) => (
+          <img
+            src="/svg/trashbin.svg"
+            style={{ width: '14px', cursor: 'pointer' }}
+            onClick={() => handleDelete(row.original._id)}
+          />
+        ),
+      },
+    ],
+    []
   )
+
+  const { getTableProps, getTableBodyProps, headerGroups, prepareRow, rows } = useTable({
+    columns,
+    data,
+  })
 
   useEffect(() => {
     if (!tableContainerRef || !tableContainerRef.current) return
